refactor(semantics): unify storage-view and type queries, remove lowering dependency (closes #914)
- Add src/semantics/storageView.ts: canonical collectModuleStorage(),
  collectFunctionLocals(), ScalarKind resolver, and shared diagAt() helper —
  single copy replacing the identical implementations duplicated in both
  acceptance files.

- Add src/semantics/typeQueries.ts: move createTypeResolutionHelpers,
  ScalarKind, and AggregateType out of the lowering layer into the semantics
  layer.  Semantics modules now have no cross-layer import to lowering.

- Shrink src/lowering/typeResolution.ts to a 16-line re-export shim so all
  existing lowering-layer imports continue to resolve without change.

- Add src/semantics/instructionAcceptance.ts: runInstructionAcceptance()
  traversal framework that collects module storage once, builds per-function
  helpers once, and dispatches to pluggable InstructionValidator callbacks.

- Refactor assignmentAcceptance.ts (249→99 lines): remove duplicated storage
  collection and lowering import; reduce to a thin InstructionValidator policy
  module using runInstructionAcceptance.

- Refactor succPredAcceptance.ts (199→47 lines): same treatment.

No behaviour changes; all tests pass.
