+++ src/semantics/assignmentAcceptance.ts
@@ -1,32 +1,12 @@
-import { DiagnosticIds } from '../diagnostics/types.js';
 import type { Diagnostic } from '../diagnostics/types.js';
-import type {
-  AsmInstructionNode,
-  EaExprNode,
-  FuncDeclNode,
-  OpDeclNode,
-  ProgramNode,
-  SourceSpan,
-  TypeExprNode,
-  VarDeclNode,
-} from '../frontend/ast.js';
-import { visitDeclTree } from './declVisitor.js';
+import type { ProgramNode } from '../frontend/ast.js';
 import type { CompileEnv } from './env.js';
-import { createTypeResolutionHelpers, type ScalarKind } from '../lowering/typeResolution.js';
-import { resolveVisibleType } from '../moduleVisibility.js';
+import { diagAt } from './storageView.js';
+import type { ScalarKind } from './typeQueries.js';
+import { runInstructionAcceptance } from './instructionAcceptance.js';
+import type { InstructionValidator } from './instructionAcceptance.js';
 
 type TransferCompatibility = ScalarKind | 'address';
-
-function diagAt(diagnostics: Diagnostic[], span: SourceSpan, message: string): void {
-  diagnostics.push({
-    id: DiagnosticIds.SemanticsError,
-    severity: 'error',
-    message,
-    file: span.file,
-    line: span.start.line,
-    column: span.start.column,
-  });
-}
 
 function compatibleScalarTransfer(
   dst: TransferCompatibility | undefined,
@@ -40,210 +20,80 @@ function compatibleScalarTransfer(
   );
 }
 
-function collectModuleStorage(program: ProgramNode, env: CompileEnv) {
-  const storageTypes = new Map<string, TypeExprNode>();
-  const rawAddressSymbols = new Set<string>();
-  const moduleAliasTargets = new Map<string, EaExprNode>();
+const assignmentValidator: InstructionValidator = {
+  validateInstruction(item, _storage, helpers, diagnostics) {
+    if (item.head !== ':=' || item.operands.length !== 2) return;
+    const [dst, src] = item.operands;
+    if (dst?.kind !== 'Ea' || src?.kind !== 'Ea') return;
 
-  const resolveScalarKind = (typeExpr: TypeExprNode, seen: Set<string> = new Set()): ScalarKind | undefined => {
-    if (typeExpr.kind !== 'TypeName') return undefined;
-    const lower = typeExpr.name.toLowerCase();
-    if (lower === 'byte' || lower === 'word' || lower === 'addr') return lower;
-    if (lower === 'ptr') return 'addr';
-    if (seen.has(lower)) return undefined;
-    seen.add(lower);
-    const decl = resolveVisibleType(typeExpr.name, typeExpr.span.file, env);
-    if (!decl || decl.kind !== 'TypeDecl') return undefined;
-    return resolveScalarKind(decl.typeExpr, seen);
-  };
+    const dstType = helpers.resolveEaTypeExpr(dst.expr);
+    const dstScalar = helpers.resolveScalarTypeForLd(dst.expr);
 
-  visitDeclTree(program.files.flatMap((file) => file.items), (item, ctx) => {
-    const namedSection = ctx.section;
-    switch (item.kind) {
-      case 'VarBlock':
-        if (item.scope !== 'module' || namedSection) return;
-        for (const decl of item.decls) {
-          const lower = decl.name.toLowerCase();
-          if (decl.form === 'typed') storageTypes.set(lower, decl.typeExpr);
-          else moduleAliasTargets.set(lower, decl.initializer.expr);
-        }
-        return;
-      case 'DataBlock':
-        for (const decl of item.decls) {
-          const lower = decl.name.toLowerCase();
-          storageTypes.set(lower, decl.typeExpr);
-          if (!resolveScalarKind(decl.typeExpr)) rawAddressSymbols.add(lower);
-        }
-        return;
-      case 'DataDecl': {
-        if (namedSection && namedSection.section !== 'data') return;
-        const lower = item.name.toLowerCase();
-        storageTypes.set(lower, item.typeExpr);
-        if (!resolveScalarKind(item.typeExpr)) rawAddressSymbols.add(lower);
+    if (src.explicitAddressOf) {
+      const srcType = helpers.resolveEaTypeExpr(src.expr);
+      if (!dstScalar || (dstScalar !== 'word' && dstScalar !== 'addr')) {
+        diagAt(
+          diagnostics,
+          item.span,
+          `":=" address-of source requires a word/addr destination.`,
+        );
         return;
       }
-      case 'BinDecl':
-      case 'HexDecl':
-      case 'RawDataDecl':
-        rawAddressSymbols.add(item.name.toLowerCase());
-        return;
-      default:
-        return;
+      if (!srcType) {
+        diagAt(
+          diagnostics,
+          item.span,
+          `":=" address-of source must reference typed storage, not a raw symbol.`,
+        );
+      }
+      return;
     }
-  });
 
-  return { storageTypes, rawAddressSymbols, moduleAliasTargets };
-}
-
-function collectFunctionLocals(decls: VarDeclNode[]) {
-  const stackSlotTypes = new Map<string, TypeExprNode>();
-  const localAliasTargets = new Map<string, EaExprNode>();
-  for (const decl of decls) {
-    const lower = decl.name.toLowerCase();
-    if (decl.form === 'typed') stackSlotTypes.set(lower, decl.typeExpr);
-    else localAliasTargets.set(lower, decl.initializer.expr);
-  }
-  return { stackSlotTypes, localAliasTargets };
-}
-
-function validateAssignmentInstruction(
-  item: AsmInstructionNode,
-  env: CompileEnv,
-  storageTypes: Map<string, TypeExprNode>,
-  rawAddressSymbols: Set<string>,
-  moduleAliasTargets: Map<string, EaExprNode>,
-  stackSlotTypes: Map<string, TypeExprNode>,
-  localAliasTargets: Map<string, EaExprNode>,
-  diagnostics: Diagnostic[],
-): void {
-  if (item.head !== ':=' || item.operands.length !== 2) return;
-  const [dst, src] = item.operands;
-  if (dst?.kind !== 'Ea' || src?.kind !== 'Ea') return;
-
-  const helpers = createTypeResolutionHelpers({
-    env,
-    storageTypes,
-    stackSlotTypes,
-    rawAddressSymbols,
-    moduleAliasTargets,
-    getLocalAliasTargets: () => localAliasTargets,
-  });
-
-  const dstType = helpers.resolveEaTypeExpr(dst.expr);
-  const dstScalar = helpers.resolveScalarTypeForLd(dst.expr);
-
-  if (src.explicitAddressOf) {
     const srcType = helpers.resolveEaTypeExpr(src.expr);
-    if (!dstScalar || (dstScalar !== 'word' && dstScalar !== 'addr')) {
+    const srcScalar = helpers.resolveScalarTypeForLd(src.expr);
+
+    if (!dstScalar) {
+      const detail = dstType ? helpers.typeDisplay(dstType) : 'unknown';
+      diagAt(diagnostics, item.span, `":=" path target must resolve to scalar storage; got ${detail}.`);
+      return;
+    }
+    if (!srcScalar) {
+      const detail = srcType ? helpers.typeDisplay(srcType) : 'unknown';
       diagAt(
         diagnostics,
         item.span,
-        `":=" address-of source requires a word/addr destination.`,
+        `":=" path source must resolve to scalar storage; got ${detail}. Use "@path" for addresses.`,
       );
       return;
     }
-    if (!srcType) {
+    if (!compatibleScalarTransfer(dstScalar, srcScalar)) {
       diagAt(
         diagnostics,
         item.span,
-        `":=" address-of source must reference typed storage, not a raw symbol.`,
+        `":=" path-to-path transfer requires compatible scalar widths; got ${dstScalar} and ${srcScalar}.`,
       );
     }
-    return;
-  }
+  },
 
-  const srcType = helpers.resolveEaTypeExpr(src.expr);
-  const srcScalar = helpers.resolveScalarTypeForLd(src.expr);
-
-  if (!dstScalar) {
-    const detail = dstType ? helpers.typeDisplay(dstType) : 'unknown';
-    diagAt(diagnostics, item.span, `":=" path target must resolve to scalar storage; got ${detail}.`);
-    return;
-  }
-  if (!srcScalar) {
-    const detail = srcType ? helpers.typeDisplay(srcType) : 'unknown';
-    diagAt(
-      diagnostics,
-      item.span,
-      `":=" path source must resolve to scalar storage; got ${detail}. Use "@path" for addresses.`,
-    );
-    return;
-  }
-  if (!compatibleScalarTransfer(dstScalar, srcScalar)) {
-    diagAt(
-      diagnostics,
-      item.span,
-      `":=" path-to-path transfer requires compatible scalar widths; got ${dstScalar} and ${srcScalar}.`,
-    );
-  }
-}
-
-function validateFunctionAssignments(
-  item: FuncDeclNode,
-  env: CompileEnv,
-  storageTypes: Map<string, TypeExprNode>,
-  rawAddressSymbols: Set<string>,
-  moduleAliasTargets: Map<string, EaExprNode>,
-  diagnostics: Diagnostic[],
-): void {
-  const { stackSlotTypes, localAliasTargets } = collectFunctionLocals(item.locals.decls);
-  for (const param of item.params) {
-    stackSlotTypes.set(param.name.toLowerCase(), param.typeExpr);
-  }
-  for (const asmItem of item.asm.items) {
-    if (asmItem.kind !== 'AsmInstruction') continue;
-    validateAssignmentInstruction(
-      asmItem,
-      env,
-      storageTypes,
-      rawAddressSymbols,
-      moduleAliasTargets,
-      stackSlotTypes,
-      localAliasTargets,
-      diagnostics,
-    );
-  }
-}
-
-function validateOpAssignments(item: OpDeclNode, diagnostics: Diagnostic[]): void {
-  for (const asmItem of item.body.items) {
-    if (asmItem.kind !== 'AsmInstruction' || asmItem.head !== ':=' || asmItem.operands.length !== 2) continue;
-    const [dst, src] = asmItem.operands;
-    if (dst?.kind !== 'Ea') continue;
+  validateOpInstruction(item, _op, diagnostics) {
+    if (item.head !== ':=' || item.operands.length !== 2) return;
+    const [dst, src] = item.operands;
+    if (dst?.kind !== 'Ea') return;
     if (src?.kind === 'Ea') {
       const detail = src.explicitAddressOf ? 'address-of' : 'path-to-path';
       diagAt(
         diagnostics,
-        asmItem.span,
+        item.span,
         `":=" ${detail} storage-target forms are not supported inside ops in this slice.`,
       );
     }
-  }
-}
+  },
+};
 
 export function validateAssignmentAcceptance(
   program: ProgramNode,
   env: CompileEnv,
   diagnostics: Diagnostic[],
 ): void {
-  const { storageTypes, rawAddressSymbols, moduleAliasTargets } = collectModuleStorage(program, env);
-
-  for (const file of program.files) {
-    visitDeclTree(file.items, (item) => {
-      if (item.kind === 'FuncDecl') {
-        validateFunctionAssignments(
-          item,
-          env,
-          storageTypes,
-          rawAddressSymbols,
-          moduleAliasTargets,
-          diagnostics,
-        );
-        return;
-      }
-      if (item.kind === 'OpDecl') {
-        validateOpAssignments(item, diagnostics);
-      }
-    });
-  }
+  runInstructionAcceptance(program, env, diagnostics, assignmentValidator);
 }
